+++ src/serverEndpoints/deleteProfileInfo.ts
@@ -1,4 +1,3 @@
-import { Model } from '../utils/mongooseImport/mongooseImport';
 import { connectToDB } from '../utils/dbConfig/dbConfig';
 import { getFormDataObject } from '../utils/formValidation/getFormDataObject';
 import { IHrUserSchema } from '../utils/dbConfig/models/hrUserModel';
@@ -8,41 +7,49 @@ import { useTranslation } from '../utils/translation/useTranslation';
 
 export async function deleteProfileInfo(req: any, res: any) {
     try {
-        const translation = useTranslation('serverAction');
-        const formData = req.body.formData;
-        const formDataObject = getFormDataObject(formData);
+        if (typeof window === "undefined") {
+            const mongoose = await import('mongoose');
+            type Model<T = any> = typeof mongoose.Model<T>;
+            const translation = useTranslation('serverAction', req.body.injector);
+            const formData = req.body.formData;
+            const formDataObject = getFormDataObject(formData);
 
-        const Model = await connectToDB(DATABASES[formDataObject['databaseName']!]) as Model<
-            IHrUserSchema | ICandidateSchema
-        >;
-        if (!Model) {
-            console.log('ERROR_DELETE_PROFILE: Error with connecting to the database!');
-            return res.status(500).json({
-                errorMessage: translation("somethingWentWrong"),
-                error: true,
+            const Model = await connectToDB(DATABASES[formDataObject['databaseName']!]) as Model<
+                IHrUserSchema | ICandidateSchema
+            >;
+            if (!Model) {
+                console.log('ERROR_DELETE_PROFILE: Error with connecting to the database!');
+                return res.status(500).json({
+                    errorMessage: translation("somethingWentWrong"),
+                    error: true,
+                });
+            }
+            // check if user already exists
+            const profile = await Model.findById(formDataObject['id']);
+            const deletedProfile = await profile?.deleteOne();
+
+            if (!deletedProfile) {
+                console.log(
+                    'ERROR_DELETE_PROFILE: Error with deleting the profile from the database!',
+                );
+                return res.status(500).json({
+                    errorMessage: translation("cannotSaveChanges"),
+                    error: true,
+                });
+            }
+
+            if (DATABASES.hrUsers === DATABASES[formDataObject['databaseName']!]) {
+                res.setHeader('Set-Cookie', 'token=; Expires=Thu, 01 Jan 1970 00:00:00 GMT; HttpOnly');
+            }
+
+            return res.status(200).json({
+                successMessage: 'Changes saved',
+                success: true,
             });
         }
-        // check if user already exists
-        const profile = await Model.findById(formDataObject['id']);
-        const deletedProfile = await profile?.deleteOne();
-
-        if (!deletedProfile) {
-            console.log(
-                'ERROR_DELETE_PROFILE: Error with deleting the profile from the database!',
-            );
-            return res.status(500).json({
-                errorMessage: translation("cannotSaveChanges"),
-                error: true,
-            });
-        }
-
-        if (DATABASES.hrUsers === DATABASES[formDataObject['databaseName']!]) {
-            res.setHeader('Set-Cookie', 'token=; Expires=Thu, 01 Jan 1970 00:00:00 GMT; HttpOnly');
-        }
-
-        return res.status(200).json({
-            successMessage: 'Changes saved',
-            success: true,
+        return res.status(500).json({
+            errorMessage: 'Unexpected error occurred',
+            error: true,
         });
     } catch (error) {
         console.error('ERROR_DELETE_PROFILE_INFO:', error);
